Le  meilleur commit of my life

+++ GameReader/app/src/main/java/com/example/gamereader/Model/Game.kt
@@ -5,5 +5,5 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName ="Game")
-data class Game(@PrimaryKey(autoGenerate = true) var id: Int, @ColumnInfo(name = "name") var name: String,@ColumnInfo(name = "description") var description : String,@ColumnInfo(name = "plateform") var plateform : String,@ColumnInfo(name = "genre") var genre : String)
+data class Game(@PrimaryKey(autoGenerate = true) var id: Long, @ColumnInfo(name = "name") var name: String,@ColumnInfo(name = "description") var description : String,@ColumnInfo(name = "plateform") var plateform : String,@ColumnInfo(name = "genre") var genre : String)
